check: rename --all to --check-all-packages
With this option all SPDX packages in the SBOM are checked for possible
vulnerabilities, even if they are not related/used by the final binary.
This may lead to false positive reports and confusion. Purpose of this
option is mainly for testing and CI. Rename the generic --all option
to --check-all-packages and extend its help description to avoid any
possible confusion.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>