spdx: allow to set PackageLicenseDeclared tag based on manifest license value
SPDX-specification-2-2 3.15.1
---8<---
Purpose: List the licenses that have been declared by the authors of the package.
Any license information that does not originate from the package authors, e.g. license
information from a third party repository, should not be included in this field.
---8<---

This allows to explicitly specify license expression for a
component, project or submodule as compared to PackageLicenseConcluded,
which is based on information about licenses found in files via
SPDX-License-Identifier.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>