feat(spdx): allow to specify SPDX package name
Currently the name is derived from the directory hierarchy entries
and package type/marker. Some 3rd party scanning tools like
cve-bin-tool are performing vulnerably scan of the SBOM file based
on the SPDX package version and name. We provide the
SPDX ExternalRef: SECURITY tag with full CPE, but this is not
used.

With this change it's possible to specify the SPDX package name for e.g.
component in the sbom.yml.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>