feat(spdx): add cve-exclude-list support
This adds a possibility to add a CVE into a exclude list. If CVE is evaluated
and a package is not vulnerable, the CVE may be included in exclude list
and will not be reported by the esp-idf-sbom check. Information about
excluded CVE is specified directly in the manifest file and stored
in generated SBOM in the PackageComment SPDX tag. The value format of
the PackageComment is a free form text and the exclude list is stored in
a yaml format. This can be simply extended in the future to store more
information if needed.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>