feat(spdx): use product part of cpe as spdx package name
Currently the name of the spdx package has to be explicitly specified in
the manifest file and this has highest priority. If not present the
generated name(based on pkg type, path, hierarchy) is used. This adds
a step in-between these two and uses product part of the cpe as a package
name. So the order of finding package name is:

name from manifest -> name from cpe -> generated name

This will also allow to automatically use e.g. cve-bin-tool to scan
packages with cpe and without explicit name in the manifest.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>