feat(spdx)!: create minimal SBOM by default
Currently esp-idf-sbom is adding component/package files into the
SBOM if it cannot figure out other unique source for the package. For
example git repo URL with exact SHA. Also the default behavior is to
generate license and copyright information.

This can lead to quite big SBOM file and it takes more time to generate SBOM.
We are primary interested in security(CPEs) and the license/copyright/files
info is not required for vulnerability scan.

This changes the default behaviour:
1) --files option default value is changed from auto to rem, so no
   package files are added. Meaning FilesAnalyzed is by default False.

2) --no-file-tags changed to --file-tags. Meaning licenses and
   copyrights will be added only if explicitly requested.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>