feat: extend validate-submodule-hash to validate all manifest info
This extends !15, which added new validate command allowing to check sbom hashes
specified in .gitmodules. We should validate all information in the manifest file,
not the sbom-hash only.

We already have full manifest validation, but it was part of spdx
generation. Now it was extracted into new mft.py module and
now it can be easily used by spdx and also the validate command.

The hash manifest value is now also checked during manifest validation
and can be used for any directory, not only for submodules.

The validate-submodule-hash command name was changed to validate and
is now part of the esp-idf-sbom manifest sub-command.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>