feat: add esp-idf-sbom manifest check command
This allows to check for vulnerabilities based on information in
manifest files(.gitmodules, idf_component.yml, sbom.yml). It accepts
mix of files and directories, which should be checked and it prints
report similar to the  esp-idf-sbom check command output.

Since the "check" and "manifest check" commands are using same report
table layout, new report.py module is added, so it can be reused by both
commands.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>