fix(mft): do not validate hash if not in git
If sbom manifest contains hash variable, we try to validate
it against information in git-tree. But there might be situation,
where the git information is no longer available for the component.
For example once a managed component is exported(fetched by idf component manager),
it's no longer stored within a git working tree, but in a plain directory.
Meaning there is no SHA information which could be used to validate
the hash variable in sbom manifest. In these cases we just skip the
check.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>