fix(ci): do not run sbom-test job twice
The sbom-test job does not have any rules, so when a new branch is pushed,
this job is started. After the MR is created, this job is tarted once again.
At this point I don't see any reason/requirement to run pipelines for branches.
So let's run all the jobs in the pipeline only for merge requests and also
once the MR is merged into default branch.

Closes IDF-8298

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>