feat: use rich for logging and other cleanups
Currently the logging and rich usage is kinda mess. This was
created while we were adding new functionality and pretty tables.
There are multiple Console instances handling different tasks.
This change unifies Console and log usage. Now there is a single
instance(stdout/stderr), which is used for everything. This also
simplifies handling of the progress bar.

The following options are changed. This should have
no effect on the core functionality and we haven't
committed to any interface yet, so it should be fine.

* Removed --verbose option, it was mostly unused.
* Removed --print for create command
* Renamed --no-colors to --no-color to be in align with rich options and
  esp-idf-size
* Renamed --force-colors to --force-terminal to be in align with rich options and
  esp-idf-size
* Added --output-file for manifest check command
* Added --output-file for check command

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>