fix: make get_tree_sha to use git-ls-tree only
Currently when we look for the submodule hash to be verified against the
one recorded in manifest(.gitmodules), we are looking at submodule hash
which is checked out in the working tree. Meaning the test expects that
submodules are checked out in the working tree with correct hash as
recorded in git-tree. This might not be true. For example when new
worktree is created the submodules are not checked out. Also when a
branch is switched, submodules are not automatically updated. We
actually don't care what's in the working tree, but only if the hash
recorded in git-tree is the same as the one recorded in manifest.

This removes usage of the rev-parse in case when the path, for which we
are verifying the hash, is a git root directory. In this case we look at
the superproject's git-tree to find the commit hash for the submodule.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>