feat: add license command
This adds a new license command, which prints information
about licenses and copyrights used by the final application
binary. Only licenses and copyrights defined in components/packages
linked into the final binary are used. This information is already
available in the SPDXProject object, so we just print a summary based
on this info and also allow to print it in json format.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>