fix: remove '+' from set of valid SPDXID characters
SPDXID tag value has a requirement that it should contain
only letters, numbers, '.', and/or '-'. The tool already has the
sanitize_spdxid function, which should make sure that the generated
SPDXID identifiers are valid. Unfortunatelly the regex used by this
function also contains '+' as a valid charecter. This fixes the
regex and adds a test to validate SPDXID generated for subpackage
with '+' in its name.

Closes https://github.com/espressif/esp-idf-sbom/issues/9

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>