fix: skip unregistered components
In some cases, a component required by another component may not be
registered. This issue arises in the build system during the early
stages of new chip support. For instance, esp_phy is required by bt, and
the build succeeds because esp_phy is not actually linked or used.
However, esp-idf-sbom relies on project_description.json, which lists
this dependency. Since esp_phy is not registered, it is absent from
build_components in project_description, causing sbom creation to fail
with KeyError: 'esp_phy'. To address this, ignore components not
listed in build_components in project_description.json.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>