feat: introduce -n/--name option to enable querying NVD by package name
Currently, the NVD has a significant backlog of unprocessed CVEs. These
CVEs lack assigned CPEs and cannot be properly queried by the
esp-idf-sbom internal checker. According to the latest statement on the
NVD announcement page, they have secured an additional contract to
address this issue, and improvements are expected. They plan to clear
the backlog by the end of the year.

In the meantime, add a new --name option. This will perform another scan
based on a keyword, which is a package name from the CPE string in the
manifest file. Note that while the CVEs are listed in the NVD, they do
not have assigned CPEs. The search primarily relies on the CVE
description, which means there may and will be false positives. These false
positives should be resolved once the NVD analyzes the CVE and assigns
it CPE. Despite the potential for false positives, this option could help
identify possible vulnerabilities early.

The possible vulnerabilities found using the keyword search are reported
in a separate table as MAYBE vulnerable and do not count towards the
summary report.

A check for underlying socket exceptions has been added when querying
NVD, along with a retry mechanism to make the request more reliable.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>