feat: introduce virtual package support
In certain scenarios, it can be useful to specify a relationship or
dependency on a package that is not physically present during
compilation. For instance, some functionality used by an application
might be implemented in ROM and utilized by the application. An example
of this is mbedtls, which could be part of the ROM and used by the
application based on configuration options. In such cases, a virtual
package can help to express this dependency.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>