feat: allow conditional expressions in manifest files
Enable the use of expressions to conditionally include packages in the
SBOM. A new "if" key has been added to the manifest file, allowing
expressions based on configuration options. If the expression evaluates
to true, the package is included in the SBOM; otherwise, it is skipped.
The "if" key is ignored for project and component packages, as they must
always be included to ensure the SBOM is complete.

The expression evaluation is influenced by kconfig and supports the
following operators in descending order of precedence, which can be
modified using brackets.

- !: negation
- <, >, >=, <=: comparison
- =, !=: equal, not equal
- &&: logical and
- ||: logical or

Examples:
*  if: 'IDF_TARGET = "esp32"'
*  if: '(IDF_TARGET_ESP32 || IDF_TARGET = "esp32") && !IDF_TOOLCHAIN_GCC'
*  if: '1536 <= FREERTOS_ISR_STACKSIZE < 0xffff'

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>