feat: add a global list of excluded CVEs
This introduces a new excluded_cves.yaml file. It's a straightforward
YAML file where CVE IDs serve as keys and corresponding text as values.
It includes a list of CVEs that should be excluded from the check report
of unanalyzed CVEs in the NVD database when the --name option is
applied.

When a new unanalyzed CVE is reported, it should be assessed. If it is
not related to ESP-IDF, it should be added to this list to prevent it
from being reported again. This will help reduce false positive reports
and ensure that potential real vulnerabilities from unanalyzed CVEs are
not overlooked.

A six-second wait was added between NVD requests as recommended by their
Best Practices, and the request timeout was increased from 30 to 60
seconds. This should improve the failure rate of requests.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>