fix: ensure pyparsing usage remains compatible with version 2.2.2 or newer
Commit bdcacd2b2aba ("feat: allow conditional expressions in manifest
files") introduced a requirement for pyparsing. Unfortunately, the API
used in expr.py is not backward compatible with older versions of
pyparsing. This update resolves the issue by using the older camel case
API, allowing compatibility with pyparsing versions starting from 2.2.2.
This change also ensures compatibility with esp-idf v4.4, which requires
'pyparsing>=2.0.3,<2.4.0'. The minimum version of 2.2.2 is chosen
because it works with newer Python versions (3.10+), where
MutableMapping was removed from the collections module. Consequently,
pyparsing versions prior to 2.2.2 do not work with Python 3.10+. Refer
to cab6d8980163 ("Fix collections ABCs deprecation warning") in
the pyparsing repository for more details.

Fixes: bdcacd2b2aba ("feat: allow conditional expressions in manifest files")
Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>