fix: include missing dependencies for the SPDX project package
Currently, components that are not dependencies of other components are
added to the SPDX project package DEPENDS_ON relationships.
Unfortunately, the common components are not being handled correctly.
For instance, newlib, which is listed in common dependencies, is
required by the nvs_flash component. However, if the nvs_flash component
is not included in the project binary, the newlib component is not added
to the SPDX project package relationships. Fix this by ignoring
requirements from components that are not linked.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>