fix: properly manage input paths for the manifest license sub-command
Currently, the manifest license sub-command only accepts a single
string path, but it calls the get_manifests() function, which expects a
list of strings. This discrepancy causes incorrect path interpretation,
rendering the sub-command unusable.

Additionally, there is no need to restrict the input to a single path;
multiple paths or manifest files can be specified.

This update modifies the manifest license sub-command to accept a mix of
manifest files and directories, similar to the manifest validate
sub-command, thereby fixing the aforementioned bug.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>