fix: include all used components in the project's SPDX dependencies
The current assumption is that if a component is listed as a requirement
of another component, it will be included in the project's SPDX package
dependency tree through that linkage. If a component is not listed as a
requirement of another component, it is added as a direct SPDX package
dependency. Unfortunately, this approach does not always work. For
instance, the main component may not specify its dependencies, relying
instead on the build system's behavior to include all discovered
components. This issue is evident in examples like
examples/wifi/getting_started/station, where the main component does not
list any requirements. Meanwhile, the esp_wifi component requires
wpa_supplicant and vice versa, so the current logic assumes that
esp_wifi and wpa_supplicant will be indirectly included in the project's
SPDX package dependency tree because they are listed as requirements of
other components. However, since there is no requirement from the main
component to include esp_wifi or wpa_supplicant, these components are
not linked into the project's SPDX package dependency.

To resolve this, ensure that all components used by the project are
reachable from the project's SPDX package. If they are not, add them as
direct dependencies.

The revised logic is as follows:
1. Gather all component requirements into a `requirements set`.
2. Gather components not present in the `requirements set` into the
   `project requirements set`.
3. Create a `reachable set` by going through direct and indirect dependencies
   reachable from the `project requirements set`.
4. For each component, verify if it is in the `reachable set`. If it is
   not, add it to the `project requirements set` and include its direct
   and indirect dependencies in the `reachable set`.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>