fix: clone and fetch only the master branch
Currently, the local database is cloned using git-mirror, and updates
are performed with fetch without specifying any refs. This results in
fetching all refs from the repository, including GitHub's pull request
branches. Since our primary concern is the latest tip of the master
branch, ensure that only it is fetched.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>