fix: avoid calling git-grep if neither cpe nor keyword is present
The local database check uses git-grep to look for potential CVEs using
CPE or keywords specified in the manifest file. If neither CPE nor
keywords are available, git-grep is called without any search patterns,
leading to the error "fatal: this operation must be run in a work tree."
Avoid running git-grep when there is no pattern to search for.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>