fix: use version comparison as a fallback to CPE match strings
Currently, the vulnerability search in the local database relies solely
on CPE match strings specified under the given Match Criteria. The issue
is that the Match Criteria might not include CPEs for all affected
versions, only those already present in the CPE Dictionary.
Consequently, a scan using the local database might fail to report some
CVEs. For instance, the mbedtls CVE-2021-45451 has Match Criteria that
lacks a CPE for mbedtls version 2.8.28. To address this, do a fallback
version comparison between the version in the checked CPE and the
version ranges listed in cpeMatch.

[1] https://services.nvd.nist.gov/rest/json/cpematch/2.0?cveId=CVE-2021-45451

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>