fix: use stricter validation of CPE string binding
The current validation for CPE string binding is quite basic. We already
have comprehensive validation available in the CPE module, so we should
utilize it. This will help avoid scanning problems that might occur due
to an invalid CPE string binding value, as demonstrated in [1]. The
issue was caused by [2], where `~` was used in an unquoted CPE, which is
not allowed per NISTIR 7695.

[1] https://github.com/espressif/idf-extra-components/actions/runs/13981959950/job/39148871238
[2] https://github.com/espressif/idf-extra-components/pull/486

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>