fix: handle version comparison when version ranges are not available
Currently, when matchCriteriaId is unavailable, or if the check based on
the applicability language and matchCriteriaId indicates that the given
CPE is not vulnerable, we still proceed to check the CPE version against
the version ranges (such as versionStartExcluding, versionEndExcluding,
etc.) provided in cpeMatch. However, these ranges might not always be
available, and the checker simply skips such cpeMatch. To address
this, add a comparison between the version in the given CPE and the
version specified in the CPE provided by the criteria key within
cpeMatch.

Example of CPE match without version ranges: In this case, we extract
the 1.7.17 version from the "criteria" and compare it with the version in
the CPE that we are checking for vulnerabilities.
```
"cpeMatch": [
  {
    "vulnerable": true,
    "criteria": "cpe:2.3:a:cjson_project:cjson:1.7.17:*:*:*:*:*:*:*",
    "matchCriteriaId": "AB3583FB-DA96-458F-80C3-7A3E81564258"
  }
]
```

In the following example, at least the versionEndExcluding is provided,
so it's used for the version comparison.
```
"cpeMatch": [
  {
    "vulnerable": true,
    "criteria": "cpe:2.3:a:cjson_project:cjson:*:*:*:*:*:*:*:*",
    "versionEndExcluding": "1.7.18",
    "matchCriteriaId": "DB89DBA4-DAFF-437E-B8C7-D9096DDD5D95"
  }
]

```

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>